requirements.txt
# Do not remove the 'comments' in the lines below (they map packages to requirement extras)
beautifulsoup4==4.10.0                                                # development
catboost~=1.2.1                                                       # baselines
datasets==2.16.0                                                      # development
dateparser==1.2.0                                                     # development
faiss-cpu==1.7.4                                                      # development
fake-useragent==1.2.1                                                 # development
fsspec==2023.10.0                                                     # development
hana-ml==2.16.23041300                                                # development
holidays==0.59                                                        # development, training, inference
lxml==4.9.1                                                           # development
matplotlib>=3.8.0                                                     # development, training
mwparserfromhell==0.6.4                                               # development
openml==0.13.1                                                        # development
pandas==2.2.1                                                         # development, training, inference
psutil==5.9.8                                                         # development
pyarrow==15.0.0                                                       # development
pydantic==2.6.4                                                       # development
pynvml==11.5.0                                                        # development
pytest==8.1.1                                                         # development
pytest-cov==4.1.0                                                     # development
python-dateutil==2.8.2                                                # development, training, inference
python-magic==0.4.27                                                  # development
pytorch-lightning==2.0.1.post0                                        # development, training, inference
pyzmq==25.1.2                                                         # development, training, inference
scikit-learn==1.4.1.post1                                             # development, training, inference
seaborn>=v0.12.0                                                      # development, training
sentence-transformers==2.2.2                                          # development, training, inference
tiktoken==0.7.0                                                       # llm
torchinfo==1.8.0                                                      # development
torch~=2.0.1                                                          # development
transformers==4.36.0                                                  # development, training, inference
wikitables==0.5.5                                                     # development
xgboost>=2.0.0                                                        # baselines
yapf==0.31.0                                                          # development
tabpfn~=0.1.8                                                         # development
tabulate>=0.9.0                                                       # development
seaborn>=v0.12.0                                                      # development, training
matplotlib>=3.8.0                                                     # development, training
omegaconf                                                             # development
imbalanced-learn==0.12.3                                              # development, training
